+++ images/test-moodle/build.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+IMAGE_NAME=test-moodle
+
+podman build --tag=$IMAGE_NAME .