+++ test-moodle/shell.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+CONTAINER_NAME=test-moodle
+
+podman exec -it $CONTAINER_NAME /bin/bash